Sistema/requirements.txt
nicegui>=1.4.0
psycopg2-binary>=2.9.6
python-dateutil>=2.8.2
pydantic>=2.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
plotly>=5.14.0